+++ src/main/java/com/jscode/ryugeonwoo/repository/BoardRepository.java
@@ -5,9 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BoardRepository extends JpaRepository<Board, Integer> {
+public interface BoardRepository extends JpaRepository<Board, Long> {
 
-    //최근 작성된 순으로 조회, 최대 100개 조회
     List<Board> findTop100ByOrderByCreatedAt();
     List<Board> findTop100ByTitleContainingOrderByCreatedAt(String keyWord);
 }
